WhatsappBot/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.6
python-environ==0.4.54
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
openai~=0.27.2
python-dotenv~=1.0.0